requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
SQLAlchemy==2.0.25
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
alembic==1.13.1
celery==5.3.6
redis==5.0.4
